+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Added a method to draw the resulting rectangle of YOLOX to image.
+- Change typedef to private.
+
 ## 0.0.1
 
 Initial Release.